+++ src/rules/at-mixin-argumentless-call-parentheses/index.js
@@ -4,7 +4,8 @@ import { namespace } from "../../utils"
 export const ruleName = namespace("at-mixin-no-argumentless-call-parentheses")
 
 export const messages = utils.ruleMessages(ruleName, {
-  expected: "Unexpected parentheses in argumentless @mixin call",
+  expected: mixin => `Expected parentheses in mixin "${mixin}" call`,
+  rejected: mixin => `Unexpected parentheses in argumentless mixin "${mixin}" call`,
 })
 
 export default function (value) {
@@ -16,13 +17,15 @@ export default function (value) {
     if (!validOptions) { return }
 
     root.walkAtRules("include", mixinCall => {
-      // If is is "No parens in argumentless calls"
+      // If it is "No parens in argumentless calls"
       if (value === "never" && mixinCall.params.search(/\(\s*?\)\s*?$/) === -1) { return }
-      // If is is "Always use parens"
+      // If it is "Always use parens"
       if (value === "always" && mixinCall.params.search(/\(/) !== -1) { return }
 
+      const mixinName = /\s*(\S*?)\s*(?:\(|$)/.exec(mixinCall.params)[1]
+
       utils.report({
-        message: messages.expected,
+        message: messages[value === "never" ? "rejected" : "expected"](mixinName),
         node: mixinCall,
         result,
         ruleName,