+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# HEAD
+
+- Fixed: `at-mixin-no-argumentless-call-parentheses` messages
+
 # 1.4.2:
 
 - Fixed: false positives in inline comment detecting by `findCommentsInRaws` if a comment is the first/last in a file/line